+++ apps/examples/ringbuf/writer.c
@@ -60,7 +60,8 @@ static int ringbuf_writer_proc(
     case AM_EVT_RINGBUF_WAIT:
     case AM_EVT_RINGBUF_WRITE: {
         uint8_t *ptr = NULL;
-        int size = am_ringbuf_get_write_ptr(&g_ringbuf, &ptr, me->len);
+        int size = me->len;
+        (void)am_ringbuf_get_write_ptr(&g_ringbuf, &ptr, &size);
         if (size < me->len) {
             am_timer_arm_ticks(&me->timer_wait, /*ticks=*/1, /*interval=*/0);
             return AM_HSM_HANDLED();